+++ tests/upload/test_main.py
@@ -4,7 +4,7 @@ import httpx
 import pytest
 from playwright.sync_api import Page, expect
 
-from tests.conftest import TESTDATA_DIR
+from tests.conftest import TESTDATA_DIR, build_ui_label_regex
 
 
 @pytest.fixture
@@ -28,14 +28,14 @@ def test_upload(
 ) -> None:
     page = upload_page
     with page.expect_file_chooser() as fc_info:
-        page.locator("role=button[name='Select Files']").click()
+        page.get_by_test_id("select-files-button").click()
     file_chooser = fc_info.value
     file_chooser.set_files(str(TESTDATA_DIR / "test.csv"))
 
     expect(page.get_by_text("test.csv")).to_be_visible()
     page.screenshot(path="tests_test_main_test_index-after-select.png")
 
-    page.get_by_text("Submit").click()
+    page.get_by_test_id("submit-button").click()
     page.screenshot(path="tests_test_main_test_index-after-submit.png")
 
     expect(page.get_by_text("test.csv")).not_to_be_visible()
@@ -51,22 +51,24 @@ def test_upload(
 @pytest.mark.integration
 def test_empty_upload(upload_page: Page) -> None:
     page = upload_page
-    page.get_by_text("Submit").click()
+    page.get_by_test_id("submit-button").click()
     page.screenshot(path="tests_test_main_test_empty_after-submit.png")
-    expect(page.locator("text=No files to upload.")).to_be_visible()
+    expect(
+        page.get_by_text(build_ui_label_regex("upload.no_files_to_upload"))
+    ).to_be_visible()
 
 
 @pytest.mark.integration
 def test_remove_selected_file(upload_page: Page) -> None:
     page = upload_page
     with page.expect_file_chooser() as fc_info:
-        page.locator("role=button[name='Select Files']").click()
+        page.get_by_test_id("select-files-button").click()
     file_chooser = fc_info.value
     file_chooser.set_files(str(TESTDATA_DIR / "test.xml"))
 
     expect(page.get_by_text("test.xml")).to_be_visible()
     page.screenshot(path="tests_test_main_test_index-after-select2.png")
-    page.get_by_role("button").and_(page.get_by_title("remove file")).click()
+    page.get_by_test_id("remove-file-button").click()
     page.screenshot(path="tests_test_main_test_index-after-delete.png")
 
     expect(page.get_by_text("test.xml")).not_to_be_visible()