+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- german and english translations for all UI labels and language switcher like mex-admin
+
 ### Changes
 
 - new template https://github.com/robert-koch-institut/mex-template/releases/tag/2.0.0
+- rename the file history page, module and route to `browse`
 
 ### Deprecated